BREAKING CHANGE: Refactoring all tests and adding the refresh token

+++ backend/src/main/java/com/streaker/service/HabitServiceImpl.java
@@ -65,6 +65,9 @@ public class HabitServiceImpl implements HabitService {
 
     @Override
     public void deleteHabit(UUID habitId) {
+        if (habitRepository.findById(habitId).isEmpty()) {
+            throw new ResourceNotFoundException("Habit not found.");
+        }
         habitRepository.deleteById(habitId);
     }
